backend/requirements.txt
Flask
flask-cors
PyYAML
PyMySQL
python-dotenv
Jinja2
